Low-Level API

AcqSynth.clear_setupdone_bit — Method

clear_setupdone_bit(boardnum)

Clear the setupDone flag, forcing the following call to setup_board to perform a full calibration.

Examples

julia> boardnum = 2
julia> clear_setupdone_bit(boardnum)

AcqSynth.get_ECL_trigger — Method

get_ECL_trigger(boardnum)

Get ECL trigger delay state. Returns 1 if the ECL trigger mode is enabled and 0 otherwise.

Function valid only for AD12 and AD8 boards.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_ECL_trigger(boardnum,4))
0

AcqSynth.get_ECL_trigger_delay — Method

get_ECL_trigger_delay(boardnum)

Get ECL trigger delay value.

Function valid only for AD12 and AD8 boards.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_ECL_trigger_delay(boardnum,4))
512

AcqSynth.get_adcresolution — Method

get_adcresolution(boardnum)

Return the bit resolution of the ADC on the board.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_adcresolution(boardnum)) # in bits
12

AcqSynth.get_all_channels — Method

get_all_channels(boardnum)

Return the number of available channels on the ADC.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_all_channels(boardnum))
2

AcqSynth.get_averager — Method

get_averager(boardnum)

Return the averager capture parameters. See set_averager for details.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> (count, depth) = get_averager(boardnum)

AcqSynth.get_channels — Method

get_channels(boardnum)

Return the current channel setup. See the set_channels function for details.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> get_channels(boardnum) # channels previously set to IN0|IN1
3

AcqSynth.get_clock — Method

get_clock(boardnum)

Return the clock being used, 1 for the internal clock, and 0 for the external clock.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_clock(boardnum))
1

AcqSynth.get_decimation — Method

get_decimation(boardnum)

Return the decimation value.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_decimation(boardnum))
8

AcqSynth.get_frequency — Method

get_frequency(boardnum)

Return the effective sampling frequency in MHz. This is a rough measurement based on the PCIe reference clock.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_frequency(boardnum)) # in MHz
1998

AcqSynth.get_memsize — Method

get_memsize(boardnum)

Return the size of the digitizer's on-board memory in MiB.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_memsize(boardnum)) # in MiB
8192

AcqSynth.get_num_boards — Method

get_num_boards()

Return the number of Ultraview boards connected to the PC.

Examples

julia> get_num_boards()
1

AcqSynth.get_num_channels — Method

get_num_channels(boardnum)

Returns the number of channels configured for acquisition. Used to help decode buffer data. For example, if two channels are in use, the samples will be arraged by channel first, and then by sample index: [s1_c1, s1_c2, s2_c1, s2_c2...].

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_num_channels(boardnum))
1

AcqSynth.get_pretrigger_mem — Method

get_pretrigger_mem(boardnum)

Return the number of samples to be recorded prior to the trigger.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_pretrigger_mem(boardnum))
128

AcqSynth.get_segmented_capture — Method

get_segmented_capture(boardnum)

Return the segmented capture parameters. See set_segmented_capture for details.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> (count, depth) = get_segmented_capture(boardnum)

AcqSynth.get_serial — Method

get_serial(boardnum)

Return the serial number of the chosen board.

Note that the DLL function requires "get_usercode.svf" to be in the current directory when it is executed.

Examples

julia> boardnum = 0
julia> get_serial(boardnum)
1092

AcqSynth.get_trigger — Method

get_trigger(boardnum)

Return the trigger type. Returns 0 for NO_TRIGGER, 1 for WAVEFORM_TRIGGER, 2 for SYNC_SELECTIVE_RECORDING, 3 for HETERODYNE or 4 for TTL_TRIGGER_EDGE.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> println(get_trigger(boardnum))
1

AcqSynth.get_waveform_trigger_params — Method

get_waveform_trigger_params(boardnum)

Return the current waveform trigger threshold and hysteresis values.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> (threshold, hysteresis) = get_waveform_trigger_params(boardnum)
julia> println("treshold: $threshold, hysteresis: $hysteresis")
treshold: 2048, hysteresis: 512

AcqSynth.has_microsynth — Method

has_microsynth(boardnum)

Return true if the board has a microsynth, false otherwise. If the board does have a microsynth, it can be used to program the frequency of the ADC.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> has_microsynth(boardnum)
false

AcqSynth.is_AD12 — Method

is_AD12(boardnum)

Return true if the board is a model AD12, false otherwise.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> is_AD12(boardnum)
true

AcqSynth.is_AD14 — Method

is_AD14(boardnum)

Return true if the board is a model AD14, false otherwise.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> is_AD14(boardnum)
false

AcqSynth.is_AD16 — Method

is_AD16(boardnum)

Return true if the board is a model AD16, false otherwise.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> is_AD16(boardnum)
false

AcqSynth.mem_alloc — Method

mem_alloc()

Allocate a 1 MiB (2^20 bytes) block of DMA page aligned memory.

Examples

julia> block = mem_alloc()

AcqSynth.mem_free — Method

mem_free(block)

Free the memory block allocated by mem_alloc. Note that the variable used to reference to the block should be assigned to something else for safety.

Examples

julia> mem_free(block)
julia> block = 0
0

AcqSynth.mem_read — Method

mem_read(boardnum, block)

Read 1 block of data into a buffer previously allocated by the mem_alloc function and clears it from the board memory. Call mem_read repeatedly to transfer all the blocks that were acquired.

Examples

julia> boardnum = 2
julia> block = mem_alloc()
julia> mem_read(boardnum,block)

AcqSynth.set_ECL_trigger — Method

set_ECL_trigger(boardnum, state)

Enable (or disable) ECL trigger mode. Setting state to 1 enables the ECL trigger, 0 disables it. The board will be forced into reset, awaiting an external trigger.

Function valid only for AD12 and AD8 boards.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> set_ECL_trigger(boardnum,1)

AcqSynth.set_ECL_trigger_delay — Method

set_ECL_trigger_delay(boardnum, delay)

Sets a delay after the ECL trigger goes from zero to one before data begins acquiring. delay should be the delay in DCLK cycles (ADC outputs data in four 12-bit words/DCLK cycle). In single channel mode there are 4 samples per DCLK. In two channel mode, there are 2 samples per DCLK.

Function valid only for AD12 and AD8 boards.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> set_ECL_trigger_delay(boardnum,4)

AcqSynth.set_averager — Method

set_averager(boardnum, count, depth)

Configure the board for averaging operation. Note that segmented capture will be disabled. Set count to 0 to disable averaging. Enabling the averager changes the output data format to 32-bit samples.

Note that this function requires the board to have been set up with the setup_board function.

Arguments

boardnum::Integer: index of installed board

count::Integer: number (0 to 2^16-1) of segments to average over. Each segment's starting amplitude is determined by the currently configured trigger. Set count to 0 to disable averaging or to 1 for flow through.

depth::Integer: number of samples to acquire in each segment, must be a multiple of 16.

Examples

julia> boardnum = 2
julia> count = 2 # number of segments to average over
julia> depth = 1100 # number of samples per segment
julia> set_averager(boardnum,count,depth)

AcqSynth.set_channels — Method

set_channels(boardnum, chan_select)

Configure which channels to use for acquisition.

For AD14 or AD16 boards, chan_select should be the bitwise OR of the desired channels IN0, IN1, IN2, and IN3.

For AD12 boards, select either the bitwise OR of the desired channels AIN0 and AIN1, or DESCLKIQ for DESCLKIQ mode, or DESIQ for DESIQ mode.

In DESCLKIQ mode, the I- and Q- inputs remain electrically separate, increasing input bandwidth. In DESIQ, the I- and Q- inputs are shorted together. In either of these modes, both inputs must be externally driven.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> chan_select = IN0|IN1
julia> set_channels(boardnum,chan_select)

AcqSynth.set_clock — Method

set_clock(boardnum,chan_select)

Configure whether to use the CLOCK_INTERNAL (1), or CLOCK_EXTERNAL (0).

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> set_clock(boardnum,CLOCK_INTERNAL)

AcqSynth.set_decimation — Method

set_decimation(boardnum, deci_value)

Set ADC decimation (return a sample every deci_value samples only). deci_value should be 1, 2, 4 or 8. A value of 1 disables decimation.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> set_decimation(boardnum,8)

AcqSynth.set_pretrigger_mem — Method

set_pretrigger_mem(boardnum, samples)

Set the number of samples to be recorded prior to the trigger. Can be between 0 and 4095.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> set_pretrigger_mem(boardnum,500)

AcqSynth.set_segmented_capture — Method

set_segmented_capture(boardnum, count, depth)

Configure the board for segmented capture operation. Note that averaging will be disabled. Set count to 0 to disable segmented capture.

Note that this function requires the board to have been set up with the setup_board function.

Arguments

boardnum::Integer: index of installed board

count::Integer: number (0 to 2^32-1) of segments to acquire. Each segment's starting amplitude is determined by the currently configured trigger. Set count to 0 to disable segmented capture.

depth::Integer: number (0 to 2^32-1) of samples to acquire in each segment. Note that the last segment will have as many samples as can fit in the rest of the buffer. Must be a multiple of 16.

Examples

julia> boardnum = 2
julia> count = 2 # number of segments to acquire
julia> depth = 1100 # number of samples per segment
julia> set_segmented_capture(boardnum,count,depth)

AcqSynth.set_trigger — Function

set_trigger(boardnum, ttype, slope, channel)

Set the trigger type. If a waveform trigger is desired, it must be configured prealably with the [configure_waveform_trigger](@ref) function.

Note that this function requires the board to have been set up with the setup_board function.

Arguments

boardnum::Integer: index of installed board

ttype::Integer: trigger type, choose between NO_TRIGGER (0), WAVEFORM_TRIGGER (1), SYNC_SELECTIVE_RECORDING (2), HETERODYNE (3) or TTL_TRIGGER_EDGE (4).

slope::Integer: select whether to trigger on a FALLING_EDGE (0) or a RISING_EDGE (1).

channel::Int: channel to trigger on for analog triggering

Examples

julia> boardnum = 2
julia> ttype = 4 # TTL trigger
julia> slope = 1 # rising edge
julia> set_trigger(boardnum,ttype,slope)

AcqSynth.set_waveform_trigger_params — Function

set_waveform_trigger_params(boardnum, threshold, hysteresis=256)

Configure waveform triggering parameters. Do not forget to enable triggering with the set_trigger function.

Note that this function requires the board to have been set up with the setup_board function.

Arguments

boardnum::Integer: index of installed board

threshold::Integer: trigger treshold, choose a value between 0 and your ADC maximum value (eg, 2^16-1 for a 16-bit board).

hysteresis::Integer: this value should typically be a few times the noise amplitude of your signal, choose a value between 0 and your ADC maximum value (eg, 2^16-1 for a 16-bit board).

Examples

julia> boardnum = 2
julia> threshold = 2^11 # middle value for a 12-bit board
julia> hysteresis = 512 # estimated noise level * 3
julia> set_waveform_trigger_params(boardnum,threshold,hysteresis)
julia> ttype = 1 # waveform trigger
julia> slope = 1 # rising edge
julia> channel = IN0 # channel IN0
julia> set_trigger(boardnum,ttype,slope,channel)

AcqSynth.setup_acquire — Function

setup_acquire(boardnum, numblocks=1)

Prime the board to acquire numblocks blocks and wait for trigger. If no trigger was set up, the board will start acquiring immediately.

Note that this function requires the board to have been set up with the setup_board function.

Examples

julia> boardnum = 2
julia> numblocks = 32 # acquire 32 1MB blocks
julia> setup_acquire(boardnum,numblocks)

AcqSynth.setup_board — Method

setup_board(boardnum)

Read from ultra_config.dat, initialize and calibrate the board. Should be called to set the board to a known state at the beginning of an application.

Note: Call clear_setupdone_bit before calling setup_board to force a full calibration and setup. The first time the board is operated, the setupDone bit is false, and the board undergoes a complete calibration when setup_board is called. Subsequently, only mimimal configuration is performed.

Also note that the DLL function requires the ultra_config.dat file to be in the working directory when it is executed.